requirements.txt
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
blinker==1.9.0
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pip-tools==7.4.1
platformdirs==4.3.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyngrok==7.2.3
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
python-dotenv==1.1.0
python-telegram-bot==22.0
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
service-identity==24.2.0
setuptools==78.1.0
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
